Move nibble2base_ssse3() SIMD code to new simd.c source file
This function is always called via a function pointer, hence it is
never inlined so need not be in a header.

[Commit marked as authored by RV so git blame attributes the moved
code to Ruben, as it is unchanged from the original code. Temporarily
include simd.c from the header so this commit builds and runs. -JM]
